complete relay api

+++ models/system.go
@@ -1,7 +1,7 @@
 package models
 
 type System struct {
-	ID        uint64 `json:"id" gorm:"primarykey"`
-	LatestTaskFeeLogID uint64 `json:"latest_task_fee_log_id"`
+	ID                        uint   `json:"id" gorm:"primarykey"`
+	LatestTaskFeeLogID        uint   `json:"latest_task_fee_log_id"`
 	LatestTaskFeeLogTimestamp uint64 `json:"latest_task_fee_log_timestamp"`
 }
